Python_2/6_web_dev_backend/learning6app/app_api/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
passlib[bcrypt]
python-jose[cryptography]
email-validator
alembic
python-dotenv


